+++ src/main/scala/io/viash/ViashExport.scala
@@ -24,11 +24,10 @@ import io.circe.yaml.{Printer => YamlPrinter}
 import io.circe.syntax.EncoderOps
 import io.viash.helpers.circe._
 import java.nio.file.{Path, Paths, Files}
-import io.viash.schemas.{CollectedSchemas, JsonSchema}
+import io.viash.schemas._
 import io.circe.Json
-import io.viash.schemas.AutoComplete
 
-object ViashExport {
+object ViashExport extends Logging {
   def exportCLISchema(output: Option[Path], format: String): Unit = {
     val cli = new CLIConf(Nil)
     val data = cli.getRegisteredCommands().asJson
@@ -36,17 +35,22 @@ object ViashExport {
     if (output.isDefined) {
       Files.write(output.get, str.getBytes())
     } else {
-      println(str)
+      infoOut(str)
     }
   }
 
-  def exportAutocomplete(output: Option[Path]): Unit = {
+  def exportAutocomplete(output: Option[Path], format: String): Unit = {
     val cli = new CLIConf(Nil)
-    val str = AutoComplete.generateForBash(cli)
+    val str = 
+      format match {
+        case "bash" => AutoCompleteBash.generate(cli)
+        case "zsh" => AutoCompleteZsh.generate(cli)
+        case _ => throw new IllegalArgumentException("'format' must be either 'bash' or 'zsh'.")
+      }
     if (output.isDefined) {
       Files.write(output.get, str.getBytes())
     } else {
-      println(str)
+      infoOut(str)
     }
   }
 
@@ -56,7 +60,7 @@ object ViashExport {
     if (output.isDefined) {
       Files.write(output.get, str.getBytes())
     } else {
-      println(str)
+      infoOut(str)
     }
   }
 
@@ -67,7 +71,7 @@ object ViashExport {
     if (output.isDefined) {
       Files.write(output.get, str.getBytes())
     } else {
-      println(str)
+      infoOut(str)
     }
   }
 
@@ -77,7 +81,7 @@ object ViashExport {
     if (output.isDefined) {
       Files.write(output.get, str.getBytes())
     } else {
-      println(str)
+      infoOut(str)
     }
   }
 }