+++ src/main/scala/io/viash/functionality/Functionality.scala
@@ -287,109 +287,28 @@ case class Functionality(
   @internalFunctionality
   set_wd_to_resources_dir: Boolean = false
 ) {
-  // START OF REMOVED PARAMETERS THAT ARE STILL DOCUMENTED
-  @description("Adds the resources directory to the PATH variable when set to true. This is set to false by default.")
-  @since("Viash 0.5.5")
-  @removed("Extending the PATH turned out to be not desirable.", "", "0.5.11")
-  private val add_resources_to_path: Boolean = false
-
-  @description("One or more Bash/R/Python scripts to be used to test the component behaviour when `viash test` is invoked. Additional files of type `file` will be made available only during testing. Each test script should expect no command-line inputs, be platform-independent, and return an exit code >0 when unexpected behaviour occurs during testing.")
-  @removed("Use `test_resources` instead. No functional difference.", "0.5.13", "0.7.0")
-  private val tests: List[Resource] = Nil
-
-  @description("Setting this to false with disable this component when using namespaces.")
-  @since("Viash 0.5.13")
-  @removed("Use `status` instead.", "0.6.0", "0.7.0")
-  private val enabled: Boolean = true
-
-  @description("A list of input arguments in addition to the `arguments` list. Any arguments specified here will have their `type` set to `file` and the `direction` set to `input` by default.")
-  @example(
-    """inputs:
-      |  - name: input_file
-      |  - name: another_input""".stripMargin,
-      "yaml")
-  @exampleWithDescription(
-    """component_with_inputs
-      |  
-      |  Inputs:
-      |      input_file
-      |          type: file
-      |  
-      |      another_input
-      |          type: file""".stripMargin,
-      "bash",
-      "This results in the following output when calling the component with the `--help` argument:")
-  @since("Viash 0.5.11")
-  @removed("Use `arguments` instead.", "0.6.0", "0.7.0")
-  private val inputs: List[Argument[_]] = Nil
-
-  @description("A list of output arguments in addition to the `arguments` list. Any arguments specified here will have their `type` set to `file` and thr `direction` set to `output` by default.")
-  @example(
-    """outputs:
-      |  - name: output_file
-      |  - name: another_output""".stripMargin,
-      "yaml")
-  @exampleWithDescription(
-    """component_with_outputs
-      |  
-      |  Outputs:
-      |      output_file
-      |          type: file, output
-      |  
-      |      another_output
-      |          type: file, output""".stripMargin,
-      "bash",
-      "This results in the following output when calling the component with the `--help` argument:")
-  @since("Viash 0.5.11")
-  @removed("Use `arguments` instead.", "0.6.0", "0.7.0")
-  private val outputs: List[Argument[_]] = Nil
-  // END OF REMOVED PARAMETERS THAT ARE STILL DOCUMENTED
 
   // Combine inputs, outputs and arguments into one combined list
   def allArguments = arguments ::: argument_groups.flatMap(arg => arg.arguments)
 
-  private def addToArgGroup(argumentGroups: List[ArgumentGroup], name: String, arguments: List[Argument[_]]): Option[ArgumentGroup] = {
-    // Check whether an argument group of 'name' exists.
-    val existing = argumentGroups.find(gr => name == gr.name)
-
-    // if there are no arguments missing from the argument group, just return the existing group (if any)
+  def allArgumentGroups: List[ArgumentGroup] = {
     if (arguments.isEmpty) {
-      existing
-
-    // if there are missing arguments and there is an existing group, add the missing arguments to it
-    } else if (existing.isDefined) {
-      Some(existing.get.copy(
-        arguments = existing.get.arguments.toList ::: arguments
-      ))
-    
-    // else create a new group
+      // if there are no arguments, just return the argument groups as is
+      argument_groups
+    } else if (argument_groups.exists(_.name == "Arguments")) {
+      // if there is already an argument group named 'Arguments', extend it with the arguments
+      argument_groups.map{
+        case gr if gr.name == "Arguments" =>
+          gr.copy(arguments = gr.arguments ::: arguments)
+        case gr => gr
+      }
     } else {
-      Some(ArgumentGroup(
-        name = name,
+      // else create a new argument group
+      argument_groups ::: List(ArgumentGroup(
+        name = "Arguments",
         arguments = arguments
       ))
     }
-  }
-
-  def allArgumentGroups: List[ArgumentGroup] = {
-    val joinGroup = Map("Inputs" -> inputs, "Outputs" -> outputs, "Arguments" -> arguments)
-    val groups = argument_groups.map{gr => 
-      if (List("Inputs", "Outputs", "Arguments") contains (gr.name)) {
-        addToArgGroup(argument_groups, gr.name, joinGroup(gr.name)).get
-      } else {
-        gr
-      }
-    }
-    val missingGroups = joinGroup.flatMap{
-      case (name, args) =>
-        if (!groups.exists(_.name == name)) {
-          addToArgGroup(argument_groups, name, args)
-        } else {
-          None
-        }
-    }.toList
-
-    missingGroups ::: groups
   }
     
   // check whether there are not multiple positional arguments with multiplicity >1