+++ src/test/scala/io/viash/auxiliary/MainBuildAuxiliaryDockerResourceCopying.scala
@@ -2,13 +2,14 @@ package io.viash.auxiliary
 
 import io.viash.{DockerTest, TestHelper}
 import io.viash.config.Config
-import io.viash.helpers.{IO, Exec}
+import io.viash.helpers.{IO, Exec, Logger}
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.funsuite.AnyFunSuite
 
 import java.nio.file.{Files, Paths, StandardCopyOption}
 
 class MainBuildAuxiliaryDockerResourceCopying extends AnyFunSuite with BeforeAndAfterAll {
+  Logger.UseColorOverride.value = Some(false)
   private val temporaryFolder = IO.makeTemp("viash_tester")
   private val tempFolStr = temporaryFolder.toString