+++ src/test/scala/io/viash/e2e/test/MainTestNativeSuite.scala
@@ -4,7 +4,7 @@ import io.viash._
 
 import java.nio.file.{Files, Paths, StandardCopyOption}
 
-import io.viash.helpers.{IO, Exec}
+import io.viash.helpers.{IO, Exec, Logger}
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.funsuite.AnyFunSuite
 
@@ -12,6 +12,7 @@ import scala.reflect.io.Directory
 import sys.process._
 
 class MainTestNativeSuite extends AnyFunSuite with BeforeAndAfterAll {
+  Logger.UseColorOverride.value = Some(false)
   // default yaml
   private val configFile = getClass.getResource("/testbash/config.vsh.yaml").getPath
 
@@ -141,21 +142,6 @@ class MainTestNativeSuite extends AnyFunSuite with BeforeAndAfterAll {
     assert(testText.contains("Cleaning up temporary directory"))
 
     checkTempDirAndRemove(testText, false)
-  }
-
-  test("Check standard test output with legacy 'tests' definition") {
-    val newConfigFilePath = configDeriver.derive(
-      List(
-        """.functionality.tests := .functionality.test_resources""",
-        """del(.functionality.test_resources)"""
-      ) , "legacy")
-    val testOutput = TestHelper.testMainException2[Exception](
-      "test",
-      "-p", "native",
-      newConfigFilePath
-    )
-
-    assert(testOutput.error.contains("Error: .functionality.tests was removed: Use `test_resources` instead. No functional difference. Initially deprecated 0.5.13, removed 0.7.0."))
   }
 
   test("Check config file without 'functionality' specified") {