+++ src/test/scala/io/viash/platforms/nextflow/Vdsl3StandaloneTest.scala
@@ -9,12 +9,13 @@ import java.io.File
 import java.nio.file.{Files, Path, Paths}
 import scala.io.Source
 
-import io.viash.helpers.IO
+import io.viash.helpers.{IO, Logger}
 import io.viash.{DockerTest, NextflowTest, TestHelper}
 import io.viash.NextflowTestHelper
 import java.nio.charset.StandardCharsets
 
 class Vdsl3StandaloneTest extends AnyFunSuite with BeforeAndAfterAll {
+  Logger.UseColorOverride.value = Some(false)
   // temporary folder to work in
   private val temporaryFolder = IO.makeTemp("viash_tester_nextflowvdsl3")
   private val tempFolFile = temporaryFolder.toFile